refactor(db): Split merge_handler.rs into directory module
Break up the 2,037-line merge_handler.rs into a directory module with
5 focused submodules matching the CRDT delta fan-out pattern. Zero
behavioral changes — purely mechanical split with visibility adjustments.
