+++ crates/db/src/merge_handler/lww.rs
@@ -1,0 +1,229 @@
+use super::*;
+
+impl<S: Store, B: blockstore::Blockstore + Send + Sync> DbMergeHandler<S, B> {
+    /// Process an LWW delta from a block (standalone, with its own transaction).
+    pub(crate) async fn process_lww_delta(
+        &self,
+        cid: &Cid,
+        payload: &defra_core::block::LwwDeltaPayload,
+        _metadata: &BlockMetadata<'_>,
+    ) -> std::result::Result<MergeOutcome, MergeError> {
+        tracing::debug!(
+            cid = %cid,
+            field_name = %payload.field_name,
+            priority = payload.priority,
+            "Processing LWW delta"
+        );
+
+        // Create a new transaction for this merge
+        let txn = self.db.new_txn(false).await?;
+
+        // Create the LWW CRDT for this field
+        let lww = Lww::new(
+            payload.schema_version_id.clone(),
+            &payload.doc_id,
+            payload.field_name.clone(),
+        )
+        .map_err(|e| MergeError::MergeFailed(e.to_string()))?;
+
+        // Create the delta
+        let delta = LwwDelta::new(
+            payload.doc_id.clone(),
+            payload.field_name.clone(),
+            payload.priority,
+            payload.schema_version_id.clone(),
+            payload.data.clone(),
+        )
+        .map_err(|e| MergeError::MergeFailed(e.to_string()))?;
+
+        // Create the context
+        let doc_id_str = String::from_utf8_lossy(&payload.doc_id).to_string();
+        let ctx = Context {
+            doc_id: DocId::new(&doc_id_str),
+            schema_version: payload.schema_version_id.clone(),
+        };
+
+        // Perform the merge in a scoped block to ensure datastore reference is dropped
+        // before we try to commit/discard the transaction.
+        let result = {
+            let mut datastore = txn.datastore()?;
+            lww.merge(&mut datastore, &ctx, &delta).await
+        };
+
+        match result {
+            Ok(merge_result) => {
+                if merge_result.was_applied() {
+                    // Commit the transaction
+                    txn.force_commit().await?;
+                    tracing::info!(
+                        cid = %cid,
+                        field_name = %payload.field_name,
+                        doc_id = %doc_id_str,
+                        "LWW delta merged successfully"
+                    );
+                    Ok(MergeOutcome::Merged)
+                } else if merge_result.was_rejected() {
+                    // Discard the transaction - nothing to commit
+                    if let Err(e) = txn.force_discard() {
+                        tracing::error!(
+                            cid = %cid,
+                            error = %e,
+                            "Failed to discard transaction after CRDT rejection - potential resource leak"
+                        );
+                    }
+                    tracing::debug!(
+                        cid = %cid,
+                        field_name = %payload.field_name,
+                        "LWW delta rejected by CRDT (lower priority or tie-break)"
+                    );
+                    Ok(MergeOutcome::skipped(
+                        "rejected by CRDT conflict resolution",
+                    ))
+                } else {
+                    // Skipped (already applied)
+                    if let Err(e) = txn.force_discard() {
+                        tracing::error!(
+                            cid = %cid,
+                            error = %e,
+                            "Failed to discard transaction after skip - potential resource leak"
+                        );
+                    }
+                    tracing::debug!(
+                        cid = %cid,
+                        field_name = %payload.field_name,
+                        "LWW delta skipped (already applied)"
+                    );
+                    Ok(MergeOutcome::skipped("already applied"))
+                }
+            }
+            Err(e) => {
+                if let Err(discard_err) = txn.force_discard() {
+                    tracing::error!(
+                        cid = %cid,
+                        discard_error = %discard_err,
+                        merge_error = %e,
+                        "Failed to discard transaction after merge error - potential resource leak"
+                    );
+                }
+                Err(MergeError::MergeFailed(e.to_string()))
+            }
+        }
+    }
+
+    /// Process an LWW delta within an existing transaction, returning the merge result
+    /// and the winning value for document reconstruction.
+    pub(crate) async fn process_lww_delta_in_txn(
+        &self,
+        datastore: &mut NamespaceView,
+        cid: &Cid,
+        payload: &defra_core::block::LwwDeltaPayload,
+    ) -> std::result::Result<LwwMergeResult, MergeError> {
+        tracing::debug!(
+            cid = %cid,
+            field_name = %payload.field_name,
+            priority = payload.priority,
+            "Processing LWW delta in transaction"
+        );
+
+        // Create the LWW CRDT for this field
+        let lww = Lww::new(
+            payload.schema_version_id.clone(),
+            &payload.doc_id,
+            payload.field_name.clone(),
+        )
+        .map_err(|e| MergeError::MergeFailed(e.to_string()))?;
+
+        // Create the delta
+        let delta = LwwDelta::new(
+            payload.doc_id.clone(),
+            payload.field_name.clone(),
+            payload.priority,
+            payload.schema_version_id.clone(),
+            payload.data.clone(),
+        )
+        .map_err(|e| MergeError::MergeFailed(e.to_string()))?;
+
+        // Create the context
+        let doc_id_str = String::from_utf8_lossy(&payload.doc_id).to_string();
+        let ctx = Context {
+            doc_id: DocId::new(&doc_id_str),
+            schema_version: payload.schema_version_id.clone(),
+        };
+
+        // Perform the merge
+        let merge_result = lww
+            .merge(datastore, &ctx, &delta)
+            .await
+            .map_err(|e| MergeError::MergeFailed(e.to_string()))?;
+
+        // Determine the winning value for document reconstruction
+        let (applied, value) = if merge_result.was_applied() {
+            // Incoming value won - use it
+            tracing::debug!(
+                cid = %cid,
+                field_name = %payload.field_name,
+                "LWW delta applied - using incoming value"
+            );
+            let value = if !payload.data.is_empty() {
+                match ciborium::from_reader::<NormalValue, _>(&payload.data[..]) {
+                    Ok(v) => Some(v),
+                    Err(e) => {
+                        tracing::error!(
+                            field_name = %payload.field_name,
+                            error = %e,
+                            "Failed to decode applied field value from CBOR"
+                        );
+                        return Err(MergeError::BlockDecode(format!(
+                            "Failed to decode field '{}': {}",
+                            payload.field_name, e
+                        )));
+                    }
+                }
+            } else {
+                None // Tombstone
+            };
+            (true, value)
+        } else {
+            // Incoming value was rejected - read the winning value from CRDT storage
+            tracing::debug!(
+                cid = %cid,
+                field_name = %payload.field_name,
+                result = ?merge_result,
+                "LWW delta rejected - reading winning value from storage"
+            );
+
+            // Read the current (winning) value from storage
+            let value = match crdt::traits::ValueReader::value(&lww, datastore).await {
+                Ok(data) => {
+                    if data.is_empty() {
+                        None // Tombstone/deleted
+                    } else {
+                        match ciborium::from_reader::<NormalValue, _>(&data[..]) {
+                            Ok(v) => Some(v),
+                            Err(e) => {
+                                tracing::warn!(
+                                    field_name = %payload.field_name,
+                                    error = %e,
+                                    "Failed to decode existing field value from CBOR - skipping field"
+                                );
+                                None
+                            }
+                        }
+                    }
+                }
+                Err(e) => {
+                    // Field may not exist yet - this is not an error
+                    tracing::debug!(
+                        field_name = %payload.field_name,
+                        error = %e,
+                        "Could not read existing field value - field may not exist"
+                    );
+                    None
+                }
+            };
+            (false, value)
+        };
+
+        Ok(LwwMergeResult { applied, value })
+    }
+}
